+++ Sources/ShapeUp/Corner/CornerStyle+Animatable.swift
@@ -9,35 +9,10 @@ import SwiftUI
 
 extension CornerStyle: AnimatableProperties {
     public static var animatableProperties: some AnimatableProperty<Self> {
-        \.radius
-        \.concaveInset
-        \.relativeCorners.anchors
-        \.relativeCorners.offsets
-        \.relativeCorners.cornerStyles.radii
-        \.relativeCorners.cornerStyles.concaveInsets
-    }
-}
-
-extension Array where Element == CornerStyle {
-    var radii: [RelatableValue] {
-        get {
-            map(\.radius)
-        }
-        set {
-            update(with: newValue) { style, radius in
-                style.radius = radius
-            }
-        }
-    }
-
-    var concaveInsets: [CGFloat] {
-        get {
-            map(\.concaveInset)
-        }
-        set {
-            update(with: newValue) { style, concaveInset in
-                style.concaveInset = concaveInset
-            }
+        AnimatablePropertyGroup(id: \.name) {
+            \.radius
+            \.concaveInset
+            \.relativeCorners
         }
     }
 }